create StanzaCirculateColorGenerator class for Tree-D3
