+++ lib/ColorGenerator.js
@@ -80,26 +80,23 @@ export class CirculateColorGenerator extends ColorGenerator {
   // }
 }
 
-//Class CirculateColorGenerator for stanza
+//CirculateColorGenerator class for stanza
 export class StanzaCirculateColorGenerator extends ColorGenerator {
-  constructor(self, data, colorGroup) {
+  constructor(self, data, colorGroup, colorKey) {
     super();
     this.self = self;
     this.data = data;
     this.colorGroup = colorGroup;
-    // this.colorKey = colorKey;
+    this.colorKey = colorKey;
   }
-
-  // test() {
-  //   return this.self;
-  // }
 
   get colorGenerator() {
     const stanza = this.self;
     const data = this.data;
     const colorGroup = this.colorGroup;
-    // const colorKey = this.colorKey;
+    const colorKey = this.colorKey;
 
+    //Create an array of colors based on "togostanza-theme-series"
     const getColorSeries = () => {
       const getPropertyValue = (key) =>
         window.getComputedStyle(stanza.element).getPropertyValue(key);
@@ -112,24 +109,29 @@ export class StanzaCirculateColorGenerator extends ColorGenerator {
       }
       return { togostanzaColors: series, defaultColor: series[0] };
     };
+
+    //Create array for domain and define color scale
     const groupArray = [];
     data.forEach((d) =>
       d.data[colorGroup] ? groupArray.push(d.data[colorGroup]) : ""
     );
-
-    //この行以降を確認する
     const groupColor = d3
       .scaleOrdinal()
       .domain(groupArray)
       .range(getColorSeries().togostanzaColors.slice(1, 6));
 
-    const dataGroupColor = data.forEach((d) => groupColor(d.data[colorGroup]));
-    console.log(groupColor);
-
-    return {
-      defaultColor: getColorSeries().defaultColor,
-      dataGroup: dataGroupColor,
+    //Set color for each data
+    const setColor = (d) => {
+      if (d.data[colorKey]) {
+        return d.data[colorKey];
+      } else {
+        return d.data[colorGroup]
+          ? groupColor(d.data[colorGroup])
+          : getColorSeries().defaultColor;
+      }
     };
+
+    return setColor;
   }
 }
